requirements.txt
hydra-core
omegaconf
pytorch-lightning
torchmetrics
